Uncomment useful rspec features in spec_helper.rb
RSpec provides a number of useful features that are by default
commented out in the spec_helper configuration. This commit
uncomments all of them except the `profile_examples` option.

If the repo's test suite begins to slow down to an intolerable
degree, a later commit may uncomment profile_examples option.