Use config.ru to serve modular Sinatra application
This commit adds a CookRouletteApp class that inherits from
Sinatra::Base which gives it access to the Sinatra DSL.

Then we can serve the CookRouletteApp via a config.ru rack file.
The advantage to this is that we can serve the config.ru through
Heroku, our target server platform.

This commit follows the Sinatra README for serving a modular
application.