Replace SINATRA_ENV env var with APP_ENV
The Sinatra README refers to the environment variable that
determines the environment that Sinatra is run in as APP_ENV
where SINATRA_ENV was a constant that I made up to track the
name of the environment.

By switching names, we reduce the number of canonical sources
of the environment name.